update events logic
