+++ website/frontend/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.3",
   "private": true,
   "dependencies": {
-    "@gear-js/api": "0.19.0",
+    "@gear-js/api": "0.20.0",
     "@gear-js/ui": "0.2.5",
     "@hcaptcha/react-hcaptcha": "1.2.0",
     "@monaco-editor/react": "4.4.2",
