+++ website/frontend/src/components/pages/Explorer/children/Events/init.ts
@@ -2,7 +2,7 @@ import { Methods } from 'types/explorer';
 
 export const filterValues = {
   [Methods.TRANSFER]: false,
-  [Methods.LOG]: false,
+  [Methods.USER_MESSAGE_SENT]: false,
   [Methods.INIT_SUCCESS]: false,
   [Methods.INIT_FAILURE]: false,
   [Methods.DISPATCH_MESSAGE_ENQUEUED]: false,
