+++ website/frontend/src/hooks/useProgram.ts
@@ -21,9 +21,11 @@ const useProgram = (id?: string): [ProgramModel?, Metadata?] => {
 
   useEffect(() => {
     if (id) {
-      getProgram(id).then(({ result }) => setProgram(result)).catch((err: RPCResponseError) => alert.error(err.message));
-    } 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+      getProgram(id)
+        .then(({ result }) => setProgram(result))
+        .catch((err: RPCResponseError) => alert.error(err.message));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [id]);
 
   return [program, metadata];
