+++ website/frontend/src/types/explorer.ts
@@ -1,10 +1,10 @@
-import { GearEvent } from '@gear-js/api';
+import { GearGenericEvent } from '@gear-js/api';
 import { Compact, Vec } from '@polkadot/types';
 import { BlockNumber, Event as DotEvent } from '@polkadot/types/interfaces';
 import { FrameSystemEventRecord } from '@polkadot/types/lookup';
 import { generateRandomId } from 'helpers';
 
-export class IdeaEvent extends GearEvent {
+export class IdeaEvent extends GearGenericEvent {
   constructor(event: DotEvent, blockNumber?: Compact<BlockNumber>) {
     const { section, method, meta, hash } = event;
     const { docs } = meta;
@@ -55,11 +55,9 @@ export enum Sections {
   SYSTEM = 'system',
 }
 
-//  TODOEVENTS:
-// использовать Methods из @gear-js/api
 export enum Methods {
   TRANSFER = 'Transfer',
-  LOG = 'Log',
+  USER_MESSAGE_SENT = 'UserMessageSent',
   INIT_SUCCESS = 'InitSuccess',
   INIT_FAILURE = 'InitFailure',
   DISPATCH_MESSAGE_ENQUEUED = 'DispatchMessageEnqueued',
